social_networking_app/requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
boto3==1.34.47
botocore==1.34.47
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
jmespath==1.0.1
mysqlclient==2.2.4
openpyxl==3.1.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
